% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-presentations.R
\name{build_presentations}
\alias{build_presentations}
\title{Each Rmarkdown Xaringan presentation in \code{Presentations/} is
rendered. Rendered html and associated directories are placed within \code{docs/presentations/}}
\usage{
build_presentations(
  output_directory = "docs/presentations",
  depth = 1L,
  encoding = "UTF-8",
  quiet = TRUE,
  input_presentations_directory = "Presentations"
)
}
\arguments{
\item{output_directory}{Output path. Relative paths are taken relative to the
\code{pkg} directory.}

\item{depth}{Depth of path relative to root of documentation.  Used
to adjust relative links in the navbar.}

\item{encoding}{The encoding of the input files.}

\item{quiet}{Set to `FALSE` to display output of knitr and
pandoc. This is useful when debugging.}

\item{input_presentations_directory}{Defaults to Presentations}
}
\description{
In addition, a small file is created in \code{Documents/presentations} that is embeds the presentation
into the website generated by packagePages()/projectPages()
}
\details{
If you want articles that are not vignettes, either put them in
subdirectories or list in \code{.Rbuildignore}. An articles link
will be automatically added to the default navbar if the vignettes
directory is present: if you do not want this, you will need to
customise the navbar. See \code{\link{build_site}} details.

Build presentations
}
