% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_pdf.R
\name{tufte_pdf}
\alias{tufte_pdf}
\title{Tufte PDF function}
\usage{
tufte_pdf(
  documentclass = c("tufte-handout", "tufte-book"),
  fig_width = 4,
  fig_height = 2.5,
  fig_crop = TRUE,
  dev = "pdf",
  highlight = "default",
  template,
  latex_engine = "xelatex",
  includes,
  pandoc_args,
  ...
)
}
\arguments{
\item{documentclass}{Document Class}

\item{fig_width}{Width of figure}

\item{fig_height}{Height of figure}

\item{fig_crop}{Crop figure (default to TRUE)}

\item{dev}{Graphics device (defaults to pdf)}

\item{highlight}{Style of highlight (defaults to default)}

\item{template}{Template to use}

\item{latex_engine}{LaTeX engine to use (defaults to xelatex)}

\item{includes}{Stuff to include}

\item{pandoc_args}{Arguments passed to pandoc}

\item{...}{Further arguments to be passed in}
}
\description{
Tufte PDF function
}
